// src/types/language.ts
// Supported languages
export type LanguageCode = 
  | 'es-MX'    // Spanish (Mexico)
  | 'en-GB'    // English (UK)
  | 'en-US'    // English (USA)
  | 'es-ES'    // Spanish (Spain)
  | 'ca-ES'    // Catalan
  | 'de-AT'    // German (Austria)
  | 'de-DE'    // German (Germany)
  | 'fr-FR';   // French (France)